requirements.txt
djangorestframework-jwt
PyJWT
bcrypt
djangorestframework
asgiref==3.2.3
Django==3.0.3
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.7.7
psycopg2-binary==2.8.4